backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.19.1
flower==2.0.1
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
humanize==4.13.0
idna==3.10
jiter==0.11.0
kombu==5.5.4
llvmlite==0.44.0
numba==0.61.2
numpy==2.2.6
openai==1.109.1
packaging==25.0
pandas==2.3.2
pandas-ta==0.4.71b0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
pycoingecko==3.2.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
resend==2.15.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
